Better user API.
