+++ .gitignore
@@ -16,6 +16,11 @@ __pycache__/
 # C extensions
 *.so
 
+# images and annotations
+data/
+image_labeling/
+prediction_output/
+
 # Distribution / packaging
 .Python
 build/
