+++ sign_detection/coco_utils.py
@@ -9,7 +9,7 @@ import torchvision
 from pycocotools import mask as coco_mask
 from pycocotools.coco import COCO
 
-import transforms as T
+import transforms_custom as T
 
 
 class FilterAndRemapCocoCategories(object):
