+++ sign_detection/utils_custom.py
@@ -1,6 +1,7 @@
 import torch
 import torchvision
 from torchvision import transforms
+# from transforms_custom import Compose, ImageResize, ToTensor
 from torchvision.models.detection.faster_rcnn import FastRCNNPredictor
 
 
@@ -9,9 +10,19 @@ def get_transform(train):
     trans.append(transforms.ToTensor())
 
     if train:
-        # transforms.append(transforms.RandomHorizontalFlip(0.5))
+        transforms.append(transforms.RandomHorizontalFlip(0.5))
         pass
     return transforms.Compose(trans)
+
+# def get_transform(train):
+#    trans = []
+# 
+#    if train:
+#        # trans.append(ImageResize((512, 512)))
+#        pass
+#
+#    trans.append(ToTensor())
+#    return Compose(trans)
 
 
 def collate_fn_custom(batch):
